requirements.txt
# Framework
fastapi==0.95.2
uvicorn==0.21.1
pydantic==1.10.7
python-dotenv==1.0.0
aiohttp==3.8.4

# LINE Bot SDK
line-bot-sdk==3.1.0

# Data Processing
langchain==0.0.267
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.6.1
chromadb==0.4.15
beautifulsoup4==4.12.2
feedparser==6.0.10
pypdf==3.15.1

# Database
redis==5.0.1

# Utility
aiofiles==23.1.0
pyyaml==6.0.1
markdown==3.5.1
markdown-it-py==3.0.0
emoji==2.8.0
pillow==10.1.0
requests==2.31.0

# Testing
pytest==7.3.1
pytest-asyncio==0.21.0
httpx==0.24.0 